Updated the code

+++ section/section.c
@@ -22,9 +22,9 @@
 #include "../cvect.h"
 
 // FUNCTION TO FIND VECTOR USING SECTION FORMULA USING INTERNAL METHOD
-int *internal(int *p1, int *p2, int m, int n)
+float *internal(int *p1, int *p2, int m, int n)
 {
-    static int arr[3];
+    static float arr[3];
     for(int i=0;i<3;i++)
     {
         arr[i] = ((m*p1[i])-(n*p2[i]))/(m-n);
@@ -33,9 +33,9 @@ int *internal(int *p1, int *p2, int m, int n)
 }
 
 // FUNCTION TO FIND VECTOR USING SECTION FORMULA USING EXTERNAL METHOD
-int *external(int *p1, int *p2, int m, int n)
+float *external(int *p1, int *p2, int m, int n)
 {
-    static int arr[3];
+    static float arr[3];
     for(int i=0;i<3;i++)
     {
         arr[i] = ((m*p1[i])+(n*p2[i]))/(m+n);
